[services :: 클라이언트에서 서비스 레이어 분리]
